moved route helpers to src/utils/routeHelpers.ts
